Added tree structure functions
These do most of the work for manipulating the tree (freeing buddies, creating buddies, etc).  What still needs to be done is translate an actual memory address to a block pointer, which will be the math-iest part of the project.

After that is done, we can actually wire this into the Fops